refactor: split up DictionaryLoader into Browser and Node
use NodeDictionaryLoader as a default.
if use BrowserDictionaryLoader, build files by browserify.
