+++ src/loader/BrowserDictionaryLoader.js
@@ -1,0 +1,46 @@
+var zlib = require("zlibjs/bin/gunzip.min.js");
+var DictionaryLoader = require("./DictionaryLoader");
+/**
+ * BrowserDictionaryLoader inherits DictionaryLoader, using jQuery XHR for download
+ * @param {string} dic_path Dictionary path
+ * @constructor
+ */
+function BrowserDictionaryLoader(dic_path) {
+    DictionaryLoader.apply(this, [dic_path]);
+}
+BrowserDictionaryLoader.prototype = Object.create(DictionaryLoader.prototype);
+// BrowserDictionaryLoader.prototype.constructor = BrowserDictionaryLoader;
+
+/**
+ * Utility function to load gzipped dictionary
+ * @param {string} url Dictionary URL
+ * @param {BrowserDictionaryLoader~onLoad} callback Callback function
+ */
+BrowserDictionaryLoader.prototype.loadArrayBuffer = function (url, callback) {
+    var xhr = new XMLHttpRequest();
+    xhr.open("GET", url, true);
+    xhr.responseType = "arraybuffer";
+    xhr.onload = function () {
+        if (this.status !== 200) {
+            callback(xhr.statusText, null);
+        }
+        var arraybuffer = this.response;
+
+        var gz = new zlib.Zlib.Gunzip(new Uint8Array(arraybuffer));
+        var typed_array = gz.decompress();
+        callback(null, typed_array.buffer);
+    };
+    xhr.onerror = function (err) {
+        callback(err, null);
+    };
+    xhr.send();
+};
+
+/**
+ * Callback
+ * @callback BrowserDictionaryLoader~onLoad
+ * @param {Object} err Error object
+ * @param {Uint8Array} buffer Loaded buffer
+ */
+
+module.exports = BrowserDictionaryLoader;
